Flexible Probabilities Driven Bootstrap

This function resamples scenarios given by the flexible probabilities approach.

bootstrap_scenarios(x, p, n)

# S3 method for numeric
bootstrap_scenarios(x, p, n)

# S3 method for matrix
bootstrap_scenarios(x, p, n)

# S3 method for ts
bootstrap_scenarios(x, p, n)

# S3 method for xts
bootstrap_scenarios(x, p, n)

# S3 method for tbl
bootstrap_scenarios(x, p, n)

# S3 method for data.frame
bootstrap_scenarios(x, p, n)

Arguments

x
A time-series defining the scenario-probability distribution.

p
A vector of probabilities.

n
An integer scalar with the number of scenarios to be generated.

Value

The argument `x` is supposed to have the same size of `p`.

Examples

set.seed(123)
ret <- diff(log(EuStockMarkets))
ew  <- rep(1 / nrow(ret), nrow(ret))

bootstrap_scenarios(x = ret, p = ew, n = 10)

#>                DAX           SMI          CAC          FTSE
#>  [1,]  0.003127348 -6.111107e-03  0.003437298  0.0034884371
#>  [2,]  0.001797036  5.481834e-03 -0.005137626 -0.0007895960
#>  [3,]  0.003552768 -7.318266e-03 -0.011072032  0.0013811924
#>  [4,] -0.011119111 -3.501488e-03 -0.003436372 -0.0066650503
#>  [5,]  0.006913901  8.235236e-03  0.003848989 -0.0021594363
#>  [6,]  0.004518835  6.081200e-03  0.009384647  0.0172676371
#>  [7,]  0.003509159  3.942829e-05 -0.005271999  0.0006584827
#>  [8,]  0.003490180  3.872168e-03  0.000000000  0.0088809603
#>  [9,]  0.010007338  1.567231e-03 -0.002743068  0.0010064234
#> [10,] -0.006858512 -9.268781e-03  0.001650912  0.0112315928