Partial Information Kernel-Damping

This function uses entropy-pooling to find the probability mass that can constrain the first two moments while imposing the minimal structure in the data.

kernel_entropy(x, mean, sigma = NULL)

# S3 method for default
kernel_entropy(x, mean, sigma = NULL)

# S3 method for numeric
kernel_entropy(x, mean, sigma = NULL)

# S3 method for matrix
kernel_entropy(x, mean, sigma = NULL)

# S3 method for ts
kernel_entropy(x, mean, sigma = NULL)

# S3 method for xts
kernel_entropy(x, mean, sigma = NULL)

# S3 method for tbl_df
kernel_entropy(x, mean, sigma = NULL)

# S3 method for data.frame
kernel_entropy(x, mean, sigma = NULL)

Arguments

x
An univariate set of risk-drivers.

mean
A number in which the kernel should be centered.

sigma
A number with the uncertainty around mean. When NULL, only the mean is constrained.

Value

A tibble with the new probabilities distribution.

References

De Santis, G., R. Litterman, A. Vesval, and K. Winkelmann, 2003, Covariance matrix estimation, Modern investment management: an equilibrium approach, Wiley.

Examples

ret <- diff(log(EuStockMarkets[ , 1]))
mean <- -0.01 # scenarios around -1%
sigma <- var(diff(ret))

kernel_entropy(ret, mean, sigma)

#> # Fully-Flexible Probability 
#> # A tibble: 1,859 x 2
#>    .rowid       .p
#>     <int>    <dbl>
#>  1      1 0.000926
#>  2      2 0.000600
#>  3      3 0.000154
#>  4      4 0.000468
#>  5      5 0.000614
#>  6      6 0.000104
#>  7      7 0.000219
#>  8      8 0.000519
#>  9      9 0.000205
#> 10     10 0.000350
#> # ... with 1,849 more rows